+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/mapper/impl/ICategoryMapStructMapperImpl.java
@@ -3,8 +3,8 @@ package br.com.jpcchaves.infrastructure.mapper.impl;
 import br.com.jpcchaves.core.domain.Category;
 import br.com.jpcchaves.infrastructure.dto.category.CategoryRequestDTO;
 import br.com.jpcchaves.infrastructure.dto.category.CategoryResponseDTO;
-import br.com.jpcchaves.infrastructure.mapper.CategoryConverter;
 import br.com.jpcchaves.infrastructure.mapper.contracts.ICategoryMapper;
+import br.com.jpcchaves.infrastructure.mapper.converters.CategoryConverter;
 import br.com.jpcchaves.infrastructure.persistence.entity.CategoryEntity;
 import java.util.List;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
