+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/mapper/impl/ITodoMapStructMapperImpl.java
@@ -3,8 +3,8 @@ package br.com.jpcchaves.infrastructure.mapper.impl;
 import br.com.jpcchaves.core.domain.Todo;
 import br.com.jpcchaves.infrastructure.dto.todo.TodoRequestDTO;
 import br.com.jpcchaves.infrastructure.dto.todo.TodoResponseDTO;
-import br.com.jpcchaves.infrastructure.mapper.TodoConverter;
 import br.com.jpcchaves.infrastructure.mapper.contracts.ITodoMapper;
+import br.com.jpcchaves.infrastructure.mapper.converters.TodoConverter;
 import br.com.jpcchaves.infrastructure.persistence.entity.TodoEntity;
 import java.util.List;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
